Adding Core from v1.0
- Built out a new Error handler. Accepts custom templates and handles
both errors and exceptions more elegantly.
- Created a new validate helper class more for live unit testing
- Simplified the route class and fixed an annoying exception that
didn't make sense, when you threw an error on an construct. It
persisted throughout 1.0. Yay it's gone now.
- Ported over the same old Eden_Class file. Removed __callStatic
because it was unused on 1.0

I wanted to focus on error handling as a core instead of a second
thought. I believe having good error messaging will help reduce the
learning curve for eden faster.